HTTP client should never hang on an OAuth error
Currently, if there's an error during the OAuth phase, the HTTP client will just swallow and log the error. It doesn't callback or return a proper promise to the app which causes the http client to timeout. This could bring the node service down if there are lots of request stuck in this state.